+++ apps/client/src/app/(site)/criar-categoria/page.tsx
@@ -1,5 +1,5 @@
 import PageContainer from "@components/Container/Page";
-import CreateOrUpdateCategoryForm from "@components/forms/CreateOrUpdateCategory";
+import CreateOrUpdateCategoryForm from "@components/Forms/CreateOrUpdateCategory";
 
 export default async function CreateCategoryPage() {
   return (
