+++ apps/client/src/app/(site)/login/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import PageContainer from "@components/Container/Page";
-import LoginForm from "@components/forms/Auth/Login";
+import LoginForm from "@components/Forms/Auth/Login";
 
 export default function LoginPage() {
   return (
