+++ apps/client/src/app/(site)/oracoes/[prayerId]/editar/page.tsx
@@ -1,6 +1,6 @@
 import PageContainer from "@components/Container/Page";
 import { Heading } from "@components/Texts/Heading";
-import CreateOrUpdatePrayerForm from "@components/forms/CreateOrUpdatePrayer";
+import CreateOrUpdatePrayerForm from "@components/Forms/CreateOrUpdatePrayer";
 import { prayersProviders } from "@providers/api/prayers";
 
 interface Props {
