+++ apps/client/src/app/(site)/registrar/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import PageContainer from "@components/Container/Page";
-import LoginForm from "@components/forms/Auth/Login";
-import RegisterForm from "@components/forms/Auth/Register";
+import LoginForm from "@components/Forms/Auth/Login";
+import RegisterForm from "@components/Forms/Auth/Register";
 
 export default function LoginPage() {
   return (
